delete function implemented using forms

+++ src/Controller/KeyController.php
@@ -144,40 +144,40 @@ class KeyController extends AbstractController
 
     /**
      *
-     * @Route("/keys/delete/{id_key}", name="delete_key")
-     * @param int $id_key
+     * @Route("/keys/delete/{translation_key}", name="delete_key")
+     * @param Request $request
+     * @param TranslationKey $translation_key
      * @return Response
      */
-    public function deleteKey( int $id_key)
+    public function deleteKey(Request $request, TranslationKey $translation_key)
     {
-        //Find the selected key
-        $key_entry = $this->getDoctrine()
-            ->getRepository('App:TranslationKey')
-            ->find($id_key);
+        $form = $this->createFormBuilder($translation_key)
+            ->add('text_key')
+            ->add(
+                'translationmessages',
+                CollectionType::class,
+                array(
+                    'entry_type' => MessageType::class,
+                    'allow_add' => true,
+                )
+            )
+            ->add('submit', SubmitType::class)
+            ->getForm();
+
+        $form->handleRequest($request);
 
 
-        $data['formdata'] = $key_entry;
+        if ($form->isSubmitted()) {
+
+            $entitymanager = $this->getDoctrine()->getManager();
+            $entitymanager->remove($translation_key);
+            $entitymanager->flush();
+            return $this->redirectToRoute('index_keys');
+        } else {
+            $data['formdata'] = $form->createView();
+        }
         return $this->render('key/delete.html.twig', $data);
     }
 
-    /**
-     *
-     * @Route("/keys/delete/confim/{id_key}", name="confirm_delete")
-     * @param int $id_key
-     * @return RedirectResponse
-     */
-    public function confirmDeleteKey( int $id_key)
-    {
-        //Find the selected key
-        $translation_key = $this->getDoctrine()
-            ->getRepository('App:TranslationKey')
-            ->find($id_key);
 
-
-        $entitymanager = $this->getDoctrine()->getManager();
-        $entitymanager->remove($translation_key);
-                $entitymanager->flush();
-
-        return $this->redirectToRoute('index_keys');
-    }
 }
